src/pic/pcoderegs.c (Remove1pcode): fix warning on alpha
git-svn-id: https://sdcc.svn.sourceforge.net/svnroot/sdcc/trunk/sdcc@2509 4a8a32a2-be11-0410-ad9d-d568d2c75423

+++ src/pic/pcoderegs.c
@@ -273,7 +273,7 @@ static void Remove1pcode(pCode *pc, regs *reg, int debug_code)
     */
 
     char buff1[256];
-    int size = 256;
+    size_t size = 256;
 
     char *pbuff,**ppbuff;
     pbuff = buff1;
